slack_approval/functions/requirements-request.txt
goblet-gcp==0.8.2
slack_sdk==3.18.1
google-cloud-secret-manager==2.12.4
